Refactor protocols into modular packages
